fix: split conftest file for each respective test folder
