+++ app/backend/pytest.ini
@@ -6,3 +6,5 @@ pythonpath =
     ../..
     ../../app
     ../../app/ml
+    markers =
+        slow: marks tests as slow (deselect with '-m "not slow"')
